[USB] Use new protocol for NAND dump for du-touch S and L
See http://bluenote/duwiki/index.php/Protocole_USB#NAND_du_Dream_.28du-touch_S_et_L.29

+++ Util.h
@@ -7,7 +7,7 @@
 
 namespace Util
 {
-    QString sizeToString(int size);
+    QString sizeToString(quint64 size);
     QString timeToString(qint64 msTime);
 
     template <typename T>
